+++ src/main/java/com/github/shap_po/shappoli/integration/walkers/condition/type/entity/HasShapeAbilityEntityConditionType.java
@@ -1,7 +1,7 @@
 package com.github.shap_po.shappoli.integration.walkers.condition.type.entity;
 
 import com.github.shap_po.shappoli.integration.walkers.condition.type.ShappoliWalkersEntityConditionTypes;
-import com.github.shap_po.shappoli.integration.walkers.registry.ShappoliWalkersRegistries;
+import com.github.shap_po.shappoli.integration.walkers.registry.ShappoliWalkersShapeAbilityClassRegistry;
 import com.github.shap_po.shappoli.integration.walkers.util.WalkersUtil;
 import com.github.shap_po.shappoli.util.MiscUtil;
 import io.github.apace100.apoli.condition.ConditionConfiguration;
@@ -9,37 +9,27 @@ import io.github.apace100.apoli.condition.context.EntityConditionContext;
 import io.github.apace100.apoli.condition.type.EntityConditionType;
 import io.github.apace100.apoli.data.TypedDataObjectFactory;
 import io.github.apace100.calio.data.SerializableData;
-import io.github.apace100.calio.data.SerializableDataTypes;
-import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
-import net.minecraft.util.Identifier;
 import org.jetbrains.annotations.NotNull;
 import tocraft.walkers.ability.AbilityRegistry;
 import tocraft.walkers.ability.ShapeAbility;
 
 import java.util.List;
-import java.util.Objects;
 
 public class HasShapeAbilityEntityConditionType extends EntityConditionType {
     public static final TypedDataObjectFactory<HasShapeAbilityEntityConditionType> DATA_FACTORY = TypedDataObjectFactory.simple(
         new SerializableData()
-            .add("ability", SerializableDataTypes.IDENTIFIER, null)
-            .add("abilities", SerializableDataTypes.IDENTIFIERS, null),
+            .add("ability", ShappoliWalkersShapeAbilityClassRegistry.DATA_TYPE, null)
+            .add("abilities", ShappoliWalkersShapeAbilityClassRegistry.DATA_TYPE.list(), null),
         data -> new HasShapeAbilityEntityConditionType(MiscUtil.listFromData(data, "ability", "abilities")),
         (conditionType, serializableData) -> serializableData.instance()
             .set("abilities", conditionType.abilities)
     );
 
-    private final List<Identifier> abilities;
-    private final List<? extends Class<? extends ShapeAbility<?>>> abilityClasses;
+    private final List<? extends Class<? extends ShapeAbility<?>>> abilities;
 
-    public HasShapeAbilityEntityConditionType(List<Identifier> abilities) {
+    public HasShapeAbilityEntityConditionType(List<? extends Class<? extends ShapeAbility<?>>> abilities) {
         this.abilities = abilities;
-        this.abilityClasses = abilities
-            .stream()
-            .map(ShappoliWalkersRegistries.SHAPE_ABILITY_CLASS::get)
-            .filter(Objects::nonNull)
-            .toList();
     }
 
     @Override
@@ -59,7 +49,7 @@ public class HasShapeAbilityEntityConditionType extends EntityConditionType {
             return false;
         }
 
-        return abilityClasses.stream().anyMatch(a -> a.isInstance(shapeAbility));
+        return abilities.stream().anyMatch(a -> a.isInstance(shapeAbility));
     }
 
     @Override