+++ src/main/java/com/github/shap_po/shappoli/integration/walkers/data/ShappoliWalkersDataTypes.java
@@ -1,4 +1,5 @@
 package com.github.shap_po.shappoli.integration.walkers.data;
 
 public class ShappoliWalkersDataTypes {
+
 }